finish refactoring component files
