+++ wordingo-web/src/components/Alphabetical/alphabetical.css
