+++ wordingo-web/src/components/Category/category.js
