+++ wordingo-web/src/components/Home/home.css
