+++ wordingo-web/src/index.js
@@ -1,15 +1,13 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import "./index.css";
-import App from "./App/App";
-import { BrowserRouter as Router } from 'react-router-dom';
+import App from "./components/App/App";
+import { BrowserRouter as Router } from "react-router-dom";
 import * as serviceWorker from "./serviceWorker";
 
-
-
 ReactDOM.render(
-    <Router>
-      <App />
-    </Router>,
-    document.getElementById('root')
-  );
+  <Router>
+    <App />
+  </Router>,
+  document.getElementById("root")
+);
